+++ backend/app/users/schema.py
@@ -11,3 +11,4 @@ class User(ORMModel):
     external_id: str
     first_name: str
     last_name: str
+    has_seen_tour: bool