+++ frontend/src/components/data-products/data-product-data-output-link-popup/data-product-data-output-link-popup.component.tsx
@@ -3,8 +3,6 @@ import type { ReactNode, RefObject } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { FormModal } from '@/components/modal/form-modal/form-modal.component.tsx';
-import type { DataOutputConfiguration, DataOutputCreateFormSchema } from '@/types/data-output';
-
 import styles from './data-product-data-output-link-popup.module.scss';
 
 type Props = {
@@ -12,7 +10,7 @@ type Props = {
     isOpen: boolean;
     title: ReactNode;
     children: ReactNode;
-    formRef: RefObject<FormInstance<DataOutputCreateFormSchema & DataOutputConfiguration> | null>;
+    formRef: RefObject<FormInstance | null>;
 };
 
 export function DataProductDataOutputLinkPopup({ onClose, isOpen, title, formRef, children }: Props) {