+++ frontend/src/routes.tsx
@@ -5,19 +5,18 @@ import PublicLayout from '@/components/layout/public/public.layout.tsx';
 import RootLayout from '@/components/layout/root/root.layout.tsx';
 import { AuditLogs } from '@/pages/audit-logs/audit-logs.page.tsx';
 import { Logout } from '@/pages/auth/logout/logout-page.tsx';
+import { Cart } from '@/pages/cart/cart.page.tsx';
 import { DataProduct } from '@/pages/data-product/data-product.page.tsx';
 import { DataProductCreate } from '@/pages/data-product-create/data-product-create.page.tsx';
 import { DataProductEdit } from '@/pages/data-product-edit/data-product-edit.page.tsx';
 import { DataProducts } from '@/pages/data-products/data-products.page.tsx';
 import { Dataset } from '@/pages/dataset/dataset.page.tsx';
-import { DatasetCreate } from '@/pages/dataset-create/dataset-create.page.tsx';
 import { DatasetEdit } from '@/pages/dataset-edit/dataset-edit.page.tsx';
 import { ErrorRootElement } from '@/pages/error/error-root-element.page.tsx';
 import { ExplorerPage } from '@/pages/explorer/explorer.page.tsx';
 import { Home } from '@/pages/home/home.page.tsx';
 import { Marketplace } from '@/pages/marketplace/marketplace.page.tsx';
 import { ApplicationPaths } from '@/types/navigation';
-
 import ProtectedRoute from './components/layout/protected/protected.layout.tsx';
 import { DataOutput } from './pages/data-output/data-output.page.tsx';
 import { DataOutputEdit } from './pages/data-output-edit/data-output-edit.page.tsx';
@@ -77,6 +76,15 @@ const router = createBrowserRouter([
                         ],
                     },
                     {
+                        path: ApplicationPaths.MarketplaceCart,
+                        children: [
+                            {
+                                element: <Cart />,
+                                index: true,
+                            },
+                        ],
+                    },
+                    {
                         path: ApplicationPaths.Datasets,
                         children: [
                             {
@@ -86,10 +94,6 @@ const router = createBrowserRouter([
                             {
                                 path: ApplicationPaths.Dataset,
                                 element: <Dataset />,
-                            },
-                            {
-                                path: ApplicationPaths.DatasetNew,
-                                element: <DatasetCreate />,
                             },
                             {
                                 path: ApplicationPaths.DatasetEdit,